requirements.txt
altgraph==0.16.1
backports.csv==1.0.7
cairocffi==1.0.2
CairoSVG==2.3.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20181111
Django>=2.2.2
django-ckeditor==5.7.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
django-geojson==2.12.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-leaflet==0.24.0
django-localflavor==2.1
django-mathfilters==0.4.0
django-multiupload==0.5.2
django-registration==3.0
django-report-utils==0.3.15
django-rest-swagger==2.2.0
django-simple-history==2.7.2
django-tables2==2.0.6
django-uuidfield==0.5.0
django-wysiwyg==0.8.0
djangorestframework==3.9.3
et-xmlfile==1.0.1
html5lib==1.0.1
httplib2==0.12.3
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
macholib==1.11
MarkupSafe==1.1.1
oauthlib==3.0.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.6.2
Pillow==6.0.0
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.0
python-logstash==0.4.6
python3-ldap==0.9.8.4
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
reportlab==3.5.21
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
tablib==0.13.0
tinycss2==1.0.2
uritemplate==3.0.0
urllib3==1.24.3
WeasyPrint==47
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
psycopg2-binary
